backend/requirements.txt
fastapi
uvicorn
yfinance
tensorflow-cpu
keras
scikit-learn
pandas
numpy
httpx
python-dotenv
